Replace our hand-crafted checks for bounds of leds[] with isPixelValid()
Many of our old checks didn't actually test what we were using or had
off-by-ones of their own. Someday, these can become asserts or even
removed as the callers get fixed, but so much of the SM code relied on
silent failures that it's best to mimic that while we can.

+++ include/effects/matrix/PatternSMAurora.h
@@ -46,9 +46,9 @@ class PatternSMAurora : public LEDStripEffect
         byte _speed = map(speed, 1, 255, 128, 16);
 
         EVERY_N_MILLIS(80)
-        for (byte x = 0; x < LED_COLS; x++)
+        for (int x = 0; x < LED_COLS; x++)
         {
-            for (byte y = 0; y < LED_ROWS; y++)
+            for (int y = 0; y < LED_ROWS; y++)
             {
                 timer++;
                 g()->leds[XY(x, y)] = ColorFromPalette(
